cosmetic changes

+++ src/server/core/dctable.cpp
@@ -648,7 +648,7 @@ bool DCTable::saveToDatabase(DB_HANDLE hdb)
 				for(int i = 0; i < m_columns->size(); i++)
 				{
 					DCTableColumn *column = m_columns->get(i);
-               DBBind(hStmt, 2, DB_SQLTYPE_INTEGER, (INT32)(i + 1));
+               DBBind(hStmt, 2, DB_SQLTYPE_INTEGER, i + 1);
 					DBBind(hStmt, 3, DB_SQLTYPE_VARCHAR, column->getName(), DB_BIND_STATIC);
 					const SNMP_ObjectId& oid = column->getSnmpOid();
 					DBBind(hStmt, 4, DB_SQLTYPE_VARCHAR, oid.isValid() ? oid.toString().cstr() : nullptr, DB_BIND_TRANSIENT);
